separate forms features using different methods
