+++ src/app/app.component.html
@@ -8,16 +8,44 @@
     </h1>
 
     <nav class="flex justify-center gap-6 mb-6">
-      <a routerLink="/hello" class="text-blue-600 hover:underline dark:text-blue-400">Hello</a>
-      <a routerLink="/about" class="text-blue-600 hover:underline dark:text-blue-400">About</a>
-      <a routerLink="/ticket-manager" class="text-blue-600 hover:underline dark:text-blue-400">Ticketing</a>
-      <a routerLink="/lazy" class="text-blue-600 hover:underline dark:text-blue-400">Lazy</a>
-            <a *ngIf="!isLoggedIn()" routerLink="/login" class="text-blue-600 hover:underline dark:text-blue-400">Login</a>
+      <a
+        routerLink="/hello"
+        class="text-blue-600 hover:underline dark:text-blue-400"
+      >
+        Hello
+      </a>
+      <a
+        routerLink="/about"
+        class="text-blue-600 hover:underline dark:text-blue-400"
+      >
+        About
+      </a>
+      <a
+        routerLink="/ticket-manager"
+        class="text-blue-600 hover:underline dark:text-blue-400"
+      >
+        Ticketing
+      </a>
+      <a
+        routerLink="/lazy"
+        class="text-blue-600 hover:underline dark:text-blue-400"
+      >
+        Lazy
+      </a>
+      <a
+        *ngIf="!isLoggedIn()"
+        routerLink="/login"
+        class="text-blue-600 hover:underline dark:text-blue-400"
+      >
+        Login
+      </a>
       <button
         *ngIf="isLoggedIn()"
         (click)="logout()"
         class="text-blue-600 hover:underline dark:text-blue-400"
-      >Logout</button>
+      >
+        Logout
+      </button>
     </nav>
 
     <div class="text-center mb-4">
@@ -25,12 +53,17 @@
         class="px-4 py-2 rounded text-white bg-indigo-600 hover:bg-indigo-700 transition-colors duration-300"
         (click)="toggleDarkMode()"
       >
-        Toggle {{ isDarkMode() ? 'Light' : 'Dark' }} Mode
+        Toggle {{ isDarkMode() ? "Light" : "Dark" }} Mode
       </button>
     </div>
 
-    <div *ngIf="toastMessage()" class="text-center mb-4 absolute top-0 left-0 right-0">
-      <div class="inline-block px-4 py-2 rounded bg-yellow-200 text-yellow-900 transition-opacity duration-500">
+    <div
+      *ngIf="toastMessage()"
+      class="text-center mb-4 fixed top-10 left-0 right-0"
+    >
+      <div
+        class="inline-block px-4 py-2 rounded bg-yellow-200 text-yellow-900 transition-opacity duration-500"
+      >
         {{ toastMessage() }}
       </div>
     </div>
