+++ src/app/ticket-manager/ticket-form-ngrx/ticket-form-ngrx.component.scss
@@ -1,0 +1,12 @@
+.ticket-list {
+  ul {
+    list-style: none;
+    padding: 0;
+    li {
+      margin-bottom: 1rem;
+    }
+    input {
+      margin-right: 0.5rem;
+    }
+  }
+}
