+++ src/app/ticket-manager/ticket-form-signal/ticket-form-signal.component.scss
@@ -1,0 +1,9 @@
+.ticket-form {
+  margin-bottom: 2rem;
+  h3 {
+    margin-top: 1rem;
+  }
+  input {
+    margin-right: 0.5rem;
+  }
+}
