+++ src/app/ticket-manager/ticket-manager.component.html
@@ -1,138 +1,7 @@
-<div class="mt-10 border p-6 rounded-md bg-white dark:bg-gray-800 shadow">
-  <h2 class="text-2xl font-semibold mb-4 text-green-900 dark:text-green-100">
-    🎫 Ticket Manager
-  </h2>
+<div class="ticket-manager">
+  <ticket-form-signal></ticket-form-signal>
 
-  <form [formGroup]="addForm" (ngSubmit)="add()" class="mb-4 flex gap-2">
-    <input
-      formControlName="title"
-      class="flex-1 px-4 py-2 border dark:bg-gray-700 border-gray-300 dark:border-gray-600 rounded"
-      placeholder="New ticket title"
-    />
-    <button
-      type="submit"
-      class="px-4 py-2 bg-blue-600 text-white rounded hover:bg-blue-700"
-      [disabled]="addForm.invalid"
-    >
-      Add
-    </button>
-  </form>
+  <hr />
 
-  <div
-    *ngIf="addForm.get('title')?.invalid && addForm.get('title')?.touched"
-    class="text-red-500 text-sm mb-4"
-  >
-    Title is required.
-  </div>
-
-  <ul *ngIf="tickets$ | async as tickets" class="space-y-2">
-    <!-- list n edit mode in together -->
-    <li
-      *ngFor="let t of tickets"
-      class="flex justify-between items-center border p-2 rounded dark:border-gray-600"
-    >
-      <form [formGroup]="editForms[t.id]" class="flex flex-1 gap-2">
-        <input
-          formControlName="title"
-          class="flex-1 px-2 py-1 rounded dark:bg-gray-700"
-        />
-        <div class="space-x-2">
-          <button
-            *ngIf="editForms[t.id].dirty"
-            type="button"
-            (click)="update(t.id)"
-            class="px-3 py-1 bg-yellow-500 text-white rounded"
-            [disabled]="editForms[t.id].invalid"
-          >
-            Update
-          </button>
-          <button
-            *ngIf="editForms[t.id].dirty"
-            type="button"
-            (click)="revert(t.id, t.title)"
-            class="px-3 py-1 bg-gray-500 text-white rounded"
-          >
-            Revert
-          </button>
-          <button
-            type="button"
-            (click)="remove(t.id)"
-            class="px-3 py-1 bg-red-500 text-white rounded"
-          >
-            Delete
-          </button>
-        </div>
-      </form>
-    </li>
-  </ul>
-</div>
-
-<!-- 2nd version using signal -->
-<div class="p-4 max-w-xl mx-auto">
-  <h2 class="text-2xl font-bold mb-4">Ticket Manager</h2>
-
-  <!-- form mode -->
-  <div class="mb-4">
-    <input
-      type="text"
-      class="w-full px-4 py-2 border rounded shadow-sm focus:outline-none dark:bg-gray-700"
-      placeholder="Enter ticket title"
-      [(ngModel)]="newTicket"
-      name="newTicket"
-    />
-    <button
-      (click)="addTicket()"
-      class="mt-2 px-4 py-2 bg-blue-600 text-white rounded hover:bg-blue-700"
-    >
-      ➕ Add Ticket
-    </button>
-  </div>
-
-  <ul>
-    <li *ngFor="let ticket of tickets()" class="mb-2">
-      <!-- display mode -->
-      <div
-        *ngIf="editingId() !== ticket.id"
-        class="flex justify-between items-center"
-      >
-        <span>{{ ticket.title }} </span>
-        <div>
-          <button
-            (click)="editTicket(ticket.id, ticket.title)"
-            class="text-sm text-yellow-600 hover:underline mr-2"
-          >
-            ✏️ Edit
-          </button>
-          <button
-            (click)="deleteTicket(ticket.id)"
-            class="text-sm text-red-600 hover:underline"
-          >
-            🗑️ Delete
-          </button>
-        </div>
-      </div>
-
-      <!-- edit mode -->
-      <div *ngIf="editingId() === ticket.id" class="flex items-center gap-2">
-        <input
-          type="text"
-          class="px-2 py-1 border rounded dark:bg-gray-700"
-          [(ngModel)]="updatedTitle"
-          name="updatedTitle"
-        />
-        <button
-          (click)="updateTicket(ticket.id)"
-          class="px-2 py-1 bg-green-600 text-white rounded hover:bg-green-700"
-        >
-          ✅ Save
-        </button>
-        <button
-          (click)="cancelEdit()"
-          class="px-2 py-1 bg-gray-400 text-white rounded hover:bg-gray-500"
-        >
-          ❌ Cancel
-        </button>
-      </div>
-    </li>
-  </ul>
+  <ticket-form-ngrx></ticket-form-ngrx>
 </div>
