+++ src/app/ticket-manager/ticket-manager.component.scss
@@ -1,18 +1,6 @@
-form {
-  display: flex;
-  gap: 0.5rem;
-}
-
-input {
-  padding: 0.5rem;
-  border-radius: 0.375rem;
-}
-
-button {
-  transition: background-color 0.2s ease;
-}
-
-ul {
-  list-style-type: none;
-  padding: 0;
+.ticket-manager {
+  padding: 2rem;
+  h2 {
+    margin-bottom: 1rem;
+  }
 }
